+++ src/main/java/backend/zip/controller/schedule/EventController.java
@@ -1,7 +1,7 @@
-package backend.zip.controller.event;
+package backend.zip.controller.schedule;
 
 
-import backend.zip.service.event.EventService;
+import backend.zip.service.schedule.EventService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
